Add forgot password with JWT
